requirements.txt
flask
flask-sqlalchemy
python-dotenv
requests>=2.28.0,<3.0.0
Werkzeug==2.2.2
flask-cors==3.0.10
Flask-Session==0.7.0
redis>=4.0.0
hiredis==2.2.3
upstash-redis==1.2.0
setuptools>=58.0.0
packaging>=21.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary
Flask-Login
